requirements.txt
pandas
numpy
SQLAlchemy
psycopg2-binary
openpyxl
xlrd==1.2.0
apache-airflow==2.8.0
dbt-postgres==1.7.4
astronomer-cosmos==1.0.4
apache-airflow-providers-postgres==5.6.0
